requirement.txt
aiohttp==3.8.5
aiosignal==1.3.1
anyio==3.7.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
dataclasses-json==0.5.14
Django==4.2.4
djangorestframework==3.14.0
exceptiongroup==1.1.2
filelock==3.12.2
frozenlist==1.4.0
fsspec==2023.6.0
greenlet==2.0.2
grpcio==1.57.0
grpcio-tools==1.57.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.17.3
httpx==0.24.1
huggingface-hub==0.16.4
hyperframe==6.0.1
idna==3.4
Jinja2==3.1.2
joblib==1.3.2
jq==1.4.1
langchain==0.0.263
langsmith==0.0.22
MarkupSafe==2.1.3
marshmallow==3.20.1
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
nltk==3.8.1
numexpr==2.8.5
numpy==1.25.2
openai==0.27.8
openapi-schema-pydantic==1.2.4
packaging==23.1
Pillow==10.0.0
portalocker==2.7.0
protobuf==4.24.0
pydantic==1.10.12
pypdf==3.15.0
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0.1
qdrant-client==1.4.0
regex==2023.8.8
requests==2.31.0
safetensors==0.3.2
scikit-learn==1.3.0
scipy==1.11.1
semantic-version==2.10.0
sentence-transformers==2.2.2
sentencepiece==0.1.99
setuptools-rust==1.6.0
sniffio==1.3.0
SQLAlchemy==2.0.19
sqlparse==0.4.4
sympy==1.12
tenacity==8.2.2
threadpoolctl==3.2.0
tiktoken==0.4.0
tokenizers==0.13.3
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
transformers==4.31.0
typing-inspect==0.9.0
typing_extensions==4.7.1
urllib3==1.26.16
uvicorn==0.23.2
yarl==1.9.2
faiss-cpu==1.7.4
rank_bm25==0.2.2